refactor(mfa): extract BiometricsTestToolRow from TestToolMenu
TestToolMenu was calling useBiometricRegistrationStatus, useSingleExecution,
useWaitForNavigation and tracking isMFARevokeLoading unconditionally — even
when the user is logged out and the biometrics row is not rendered (it sits
inside the existing isAuthenticated guard). The biometrics-specific concerns
also bloated the menu and made unrelated reads on every render.

Move the biometrics-related JSX, state, and hook calls into a dedicated
BiometricsTestToolRow component. The new row keeps the existing trigger path
(singleExecution(waitForNavigate(() => Navigation.navigate(BIOMETRICS_TEST))))
unchanged; only the rendering location changes. TestToolMenu now renders
<BiometricsTestToolRow /> inside the same isAuthenticated block, so the
biometrics hooks run only when the row is actually shown.

Tests do not change: all mocks were module-level and continue to apply.
