+++ src/components/TestToolMenu.tsx
@@ -1,22 +1,17 @@
-import React, {useState} from 'react';
-import {Platform, View} from 'react-native';
-import useBiometricRegistrationStatus, {REGISTRATION_STATUS} from '@hooks/useBiometricRegistrationStatus';
+import React from 'react';
+import {Platform} from 'react-native';
 import useIsAuthenticated from '@hooks/useIsAuthenticated';
 import useLocalize from '@hooks/useLocalize';
 import useOnyx from '@hooks/useOnyx';
 import {useSidebarOrderedReportsActions} from '@hooks/useSidebarOrderedReports';
-import useSingleExecution from '@hooks/useSingleExecution';
 import useThemeStyles from '@hooks/useThemeStyles';
-import useWaitForNavigation from '@hooks/useWaitForNavigation';
-import {revokeMultifactorAuthenticationCredentials} from '@libs/actions/MultifactorAuthentication';
 import {isUsingStagingApi} from '@libs/ApiUtils';
-import Navigation from '@libs/Navigation/Navigation';
 import {setShouldFailAllRequests, setShouldForceOffline, setShouldSimulatePoorConnection} from '@userActions/Network';
 import {expireSessionWithDelay, invalidateAuthToken, invalidateCredentials} from '@userActions/Session';
 import {setIsDebugModeEnabled, setShouldShowBranchNameInTitle, setShouldUseStagingServer} from '@userActions/User';
 import CONFIG from '@src/CONFIG';
 import ONYXKEYS from '@src/ONYXKEYS';
-import ROUTES from '@src/ROUTES';
+import BiometricsTestToolRow from './BiometricsTestToolRow';
 import Button from './Button';
 import SoftKillTestToolRow from './SoftKillTestToolRow';
 import Switch from './Switch';
@@ -33,32 +28,9 @@ function TestToolMenu() {
     const styles = useThemeStyles();
     const {translate} = useLocalize();
     const {clearLHNCache} = useSidebarOrderedReportsActions();
-    const [isMFARevokeLoading, setIsMFARevokeLoading] = useState(false);
-    const {localCredentialID, isCurrentDeviceRegistered, otherDeviceCount, registrationStatus} = useBiometricRegistrationStatus();
-
-    const {singleExecution} = useSingleExecution();
-    const waitForNavigate = useWaitForNavigation();
-
-    /**
-     * The wrapper is needed to prevent rapid double‑taps on native from triggering multiple navigations.
-     * Context: https://github.com/Expensify/App/pull/79475#discussion_r2708230681
-     */
-    const navigateToBiometricsTestPage = singleExecution(
-        waitForNavigate(() => {
-            Navigation.navigate(ROUTES.MULTIFACTOR_AUTHENTICATION_BIOMETRICS_TEST);
-        }),
-    );
 
     // Check if the user is authenticated to show options that require authentication
     const isAuthenticated = useIsAuthenticated();
-
-    const statusTextMap = {
-        [REGISTRATION_STATUS.NEVER_REGISTERED]: translate('multifactorAuthentication.biometricsTest.statusNeverRegistered'),
-        [REGISTRATION_STATUS.NOT_REGISTERED]: translate('multifactorAuthentication.biometricsTest.statusNotRegistered'),
-        [REGISTRATION_STATUS.REGISTERED_OTHER_DEVICE]: translate('multifactorAuthentication.biometricsTest.statusRegisteredOtherDevice', {count: otherDeviceCount}),
-        [REGISTRATION_STATUS.REGISTERED_THIS_DEVICE]: translate('multifactorAuthentication.biometricsTest.statusRegisteredThisDevice'),
-    };
-    const biometricsTitle = translate('multifactorAuthentication.biometricsTest.troubleshootBiometricsStatus', {status: statusTextMap[registrationStatus]});
 
     return (
         <>
@@ -129,29 +101,8 @@ function TestToolMenu() {
                         />
                     </TestToolRow>
 
-                    {/* Allows testing the biometric multifactor authentication flow */}
-                    <TestToolRow title={biometricsTitle}>
-                        <View style={[styles.flexRow, styles.gap2]}>
-                            <Button
-                                small
-                                text={translate('multifactorAuthentication.biometricsTest.test')}
-                                onPress={() => navigateToBiometricsTestPage()}
-                            />
-                            {isCurrentDeviceRegistered && !!localCredentialID && (
-                                <Button
-                                    danger
-                                    isLoading={isMFARevokeLoading}
-                                    small
-                                    text={translate('multifactorAuthentication.revoke.revoke')}
-                                    onPress={async () => {
-                                        setIsMFARevokeLoading(true);
-                                        await revokeMultifactorAuthenticationCredentials({onlyKeyID: localCredentialID});
-                                        setIsMFARevokeLoading(false);
-                                    }}
-                                />
-                            )}
-                        </View>
-                    </TestToolRow>
+                    {/* Allows testing and revoking biometric multifactor authentication */}
+                    <BiometricsTestToolRow />
                 </>
             )}
